requirements.txt
# Core LangChain & AI
langchain>=0.3.7,<0.4.0
langchain-community>=0.3.5,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0
google-generativeai>=0.8.0,<1.0.0

# Vector Store & Database
pgvector>=0.3.0,<1.0.0
psycopg2-binary>=2.9.0,<3.0.0
sqlalchemy>=1.4,<2.0.36
asyncpg>=0.29.0,<1.0.0
supabase>=2.0.0,<3.0.0

# RAG & Embeddings
sentence-transformers>=3.0.0,<4.0.0
faiss-cpu>=1.7.0,<2.0.0
chromadb>=0.5.0,<0.6.0
rank-bm25>=0.2.0,<1.0.0

# API & Web Framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
python-multipart>=0.0.20,<1.0.0

# UI
streamlit>=1.40.0,<2.0.0
streamlit-chat>=0.1.0,<1.0.0

# Caching & Performance
redis>=5.0.0,<6.0.0
cachetools>=5.5.0,<6.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=1.0.0,<2.0.0

# Monitoring & Logging
prometheus-client>=0.21.0,<1.0.0
python-json-logger>=3.0.0,<4.0.0
loguru>=0.7.0,<1.0.0

# Utilities
tenacity>=9.0.0,<10.0.0
httpx>=0.28.0,<1.0.0
aiohttp>=3.11.0,<4.0.0
pyyaml>=6.0.0,<7.0.0
click>=8.1.0,<9.0.0
rich>=13.0.0,<14.0.0
tqdm>=4.65.0,<5.0.0

# Testing
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
pytest-cov>=6.0.0,<7.0.0
pytest-mock>=3.14.0,<4.0.0

# Development
black>=24.0.0,<25.0.0
ruff>=0.8.0,<1.0.0
mypy>=1.14.0,<2.0.0
pre-commit>=4.0.0,<5.0.0